Map dots indicate supercharger power level
This change customizes the icon to represent the curated power level
of the charging stations at a given Supercharger site.

Also commits modified SVG graphics to represent the slightly different
visual style. Given the majority of chargers today are "standard" with
120-150 kW of power, I have kept the visual red dot style as-is for
all of those chargers. I also experimented with a derivative style
I called "red_dot_standard" that had a smaller dot and a maroon color,
but decided it looked much too cluttered.

Visual (see screenshot in the pull request) styles for active/open sites are now:

- Outlined in white: the data is not curated (unknown power at site)
- Outlined in bright red: a V3 250 kW Supercharger
- Outlined in gray: an Urban (72 kW) Supercharger

Icons that have a half-dome gray overlay indicate Supercharger sites with limited operational hours.

Another approach would be to have individual layers for various charging
levels, but I thought that may be too complicated and not what users
care about, since most users will simply want to know that a charger is
known to be active and available.